requirements.txt
functions-framework==3.*
requests==2.28.1
openai
python-dotenv
firebase-admin==6.3.0
configparser
tiktoken
langchain==0.1.10
langchain-community==0.0.25
langchain-openai==0.0.8
faiss-cpu==1.8.0
pandas
google-api-python-client==2.125.0
stripe